pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Finance App Django"
authors = ["perymerdeka <tkjasikk@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.6"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.25.1"
scrapy = "^2.8.0"
scrapyd = "^1.4.1"
scrapy-djangoitem = "^1.1.1"
httpx = "^0.23.3"
beautifulsoup4 = "^4.11.2"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
